requirements_data.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=py37_0
ca-certificates=2019.10.16=0
certifi=2019.9.11=py37_0
cffi=1.12.3=py37h2e261b9_0
chardet=3.0.4=pypi_0
configargparse=0.14.0=pypi_0
configer=1.3.1=pypi_0
configparser=4.0.2=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.0=cuda10.0_0
cycler=0.10.0=pypi_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py37_1
defusedxml=0.6.0=py_0
entrypoints=0.3=py37_0
expat=2.2.6=he6710b0_0
fastcache=1.1.0=py37h7b6447c_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
freetype-py=2.1.0.post1=pypi_0
future=0.17.1=pypi_0
glib=2.56.2=hd408876_0
gmp=6.1.2=h6c8ec71_1
gmpy2=2.0.8=py37h10f8cd9_2
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
human-body-prior=0.8.5.0=pypi_0
icu=58.2=h9c2bf20_1
idna=2.9=pypi_0
imageio=2.5.0=pypi_0
imageio-ffmpeg=0.4.1=pypi_0
importlib_metadata=0.23=py37_0
intel-openmp=2019.4=243
ipykernel=5.1.2=py37h39e3cac_0
ipython=7.8.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
jedi=0.15.1=py37_0
jinja2=2.10.3=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.1.1=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.3.4=py37_0
jupyter_console=6.0.0=py37_0
jupyter_core=4.6.0=py37_0
kiwisolver=1.1.0=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.1=pypi_0
mesh-intersection=0.1.0=pypi_0
mistune=0.8.4=py37h7b6447c_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.14=py37ha843d7b_0
mkl_random=1.0.2=py37hd81dba3_0
more-itertools=7.2.0=py37_0
moviepy=1.0.2=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.1=hdf1c602_3
mpmath=1.1.0=py37_0
multipledispatch=0.6.0=pypi_0
nbconvert=5.6.0=py37_1
nbformat=4.4.0=py37_0
ncurses=6.1=he6710b0_1
networkx=2.2=pypi_0
ninja=1.9.0=py37hfd86e86_0
notebook=6.0.1=py37_0
numpy=1.18.2=pypi_0
olefile=0.46=py37_0
opencv-python=4.1.1.26=pypi_0
openssl=1.1.1d=h7b6447c_3
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.5.1=py_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pillow=6.1.0=py37h34e0f95_0
pip=19.2.2=py37_0
proglog=0.1.9=pypi_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.10=py_0
protobuf=3.9.1=pypi_0
ptyprocess=0.6.0=py37_0
pycparser=2.19=py37_0
pyglet=1.4.0b1=pypi_0
pygments=2.4.2=py_0
pyopengl=3.1.0=pypi_0
pyparsing=2.4.2=pypi_0
pyqt=5.9.2=py37h05f1152_2
pyrender=0.1.30=pypi_0
pyrr=0.10.3=pypi_0
pyrsistent=0.15.4=py37h7b6447c_0
python=3.7.4=h265db76_1
python-dateutil=2.8.0=py37_0
pytorch=1.1.0=cuda100py37he554f03_0
pywavelets=1.0.3=pypi_0
pyyaml=5.1.1=pypi_0
pyzmq=18.1.0=py37he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.5.5=py_0
readline=7.0=h7b6447c_5
requests=2.23.0=pypi_0
scikit-image=0.15.0=pypi_0
scipy=1.3.1=pypi_0
send2trash=1.5.0=py37_0
setuptools=41.0.1=py37_0
shapely=1.6.4.post2=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=py37_0
smplx=0.1.12=pypi_0
sqlite=3.29.0=h7b6447c_0
svgwrite=1.3.1=pypi_0
sympy=1.4=py37_0
tensorboardx=1.8=pypi_0
terminado=0.8.2=py37_0
testpath=0.4.2=py37_0
tk=8.6.8=hbc83047_0
torchgeometry=0.1.2=pypi_0
torchvision=0.3.0=cuda100py37h72fc40a_0
tornado=6.0.3=py37h7b6447c_0
tqdm=4.35.0=pypi_0
traitlets=4.3.3=py37_0
transforms3d=0.3.1=pypi_0
trimesh=3.2.16=pypi_0
urllib3=1.25.8=pypi_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
wheel=0.33.4=py37_0
widgetsnbextension=3.5.1=py37_0
xz=5.2.4=h14c3975_4
zeromq=4.3.1=he6710b0_3
zipp=0.6.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
